+++ src/components/process/index.tsx
@@ -2,8 +2,8 @@ import Image from "next/image"
 import Link from "next/link"
 import { useContext, useState, useEffect } from "react"
 import { ContextApplication } from "@/context"
-import {ArrowIcon} from "../../../../../public/assets/icons"
-import { DataDetails } from "../../../../../util/type";
+import {ArrowIcon} from "../../../public/assets/icons"
+import { DataDetails } from "../../../util/type";
 
 const aplicationStageStyle = (value: string) => {
